// packages/cloud-storage/src/flat/types.gen.ts
export type ErrorResponse = {
    error_code: ErrorCodeEnums;
    /**
     * A human-readable explanation specific to this occurrence of the problem.
     */
    readonly message: string;
    [key: string]: unknown;
};
// ...
export type BaseJobStatus = {
    jobId: JobId;
    requestId: RequestId;
    /**
     * The type of the asynchronous job, depends on the original API invocation that created the job.
     * Currently, the following job types are supported:
     * - file_upload
     * - file_download
     * - file_copy
     * - file_move
     * - folder_copy
     * - folder_move
     * - project_package
     *
     * ---
     * > NOTE:
     * > This list is not exhaustive and will be updated as new job types are added. Clients should expect new job types to be added in the future.
     * ---
     *
     */
    jobType: string;
};

/**
 * RunningJobStatus
 *
 * Represents the status of a running asynchronous job in the system
 */
export type RunningJobStatus = BaseJobStatus & {
    status: string;
};

/**
 * FileUploadSucceededJobStatus
 *
 * Represents the status of a file upload succeeded asynchronous job in the system
 */
export type FileUploadSucceededJobStatus = BaseJobStatus & {
    status: string;
    asset: Asset;
};

/**
 * ProjectPackageSucceededJobStatus
 *
 * Represents the status of a project package succeeded asynchronous job in the system
 */
export type ProjectPackageSucceededJobStatus = BaseJobStatus & {
    status: string;
    assets: Array<Asset>;
};

/**
 * FileDownloadSucceededJobStatus
 *
 * Represents the status of a file download succeeded asynchronous job in the system
 */
export type FileDownloadSucceededJobStatus = BaseJobStatus & {
    status: string;
    downloadResponse: FileDownloadOkResponse;
};

/**
 * FileCopySucceededJobStatus
 *
 * Represents the status of a file copy succeeded asynchronous job in the system
 */
export type FileCopySucceededJobStatus = BaseJobStatus & {
    status: string;
    asset: FileAsset;
};

/**
 * FileMoveSucceededJobStatus
 *
 * Represents the status of a file move succeeded asynchronous job in the system
 */
export type FileMoveSucceededJobStatus = BaseJobStatus & {
    status: string;
    asset: FileAsset;
};

/**
 * FolderCopySucceededJobStatus
 *
 * Represents the status of a folder copy succeeded asynchronous job in the system
 */
export type FolderCopySucceededJobStatus = BaseJobStatus & {
    status: string;
    asset: FolderAsset;
};

/**
 * FolderMoveSucceededJobStatus
 *
 * Represents the status of a folder move succeeded asynchronous job in the system
 */
export type FolderMoveSucceededJobStatus = BaseJobStatus & {
    status: string;
    asset: FolderAsset;
};

/**
 * Represents the status of a succeeded asynchronous job in the system
 *
 */
export type SucceededJobStatus = ({
    jobType: 'file_upload';
} & FileUploadSucceededJobStatus) | ({
    jobType: 'project_package';
} & ProjectPackageSucceededJobStatus) | ({
    jobType: 'file_download';
} & FileDownloadSucceededJobStatus) | ({
    jobType: 'file_copy';
} & FileCopySucceededJobStatus) | ({
    jobType: 'file_move';
} & FileMoveSucceededJobStatus) | ({
    jobType: 'folder_copy';
} & FolderCopySucceededJobStatus) | ({
    jobType: 'folder_move';
} & FolderMoveSucceededJobStatus);

/**
 * FolderCopyPartiallySucceededJobStatus
 *
 * Represents the status of a folder copy partially succeeded asynchronous job in the system
 */
export type FolderCopyPartiallySucceededJobStatus = BaseJobStatus & {
    status: string;
    asset: FolderAsset;
    /**
     * FailedAssetReferenceWithErrorResponse
     *
     * The errors that occurred during the job
     */
    errors: Array<ErrorResponse & {
        /**
         * FailedAssetReference
         */
        asset: {
            assetId: AssetId;
            /**
             * Human-readable name of the asset
             */
            name: string;
            assetType: AssetTypeEnums;
        };
    }>;
};

/**
 * ProjectPackagePartiallySucceededJobStatus
 *
 * Represents the status of a project package partially succeeded asynchronous job in the system
 */
export type ProjectPackagePartiallySucceededJobStatus = BaseJobStatus & {
    status: string;
    assets: Array<Asset>;
    /**
     * The errors that occurred during the job
     */
    errors: Array<ErrorResponse>;
};

/**
 * Represents the status of a partially succeeded asynchronous job in the system
 *
 */
export type PartiallySucceededJobStatus = ({
    jobType: 'folder_copy';
} & FolderCopyPartiallySucceededJobStatus) | ({
    jobType: 'project_package';
} & ProjectPackagePartiallySucceededJobStatus);

/**
 * FailedJobStatus
 *
 * Represents the status of a failed asynchronous job in the system
 */
export type FailedJobStatus = BaseJobStatus & {
    status: string;
    /**
     * The errors that occurred during the job
     */
    errors: Array<ErrorResponse>;
};

/**
 * JobStatus
 *
 * Represents the status of a job in the system.
 */
export type JobStatus = ({
    status: 'running';
} & RunningJobStatus) | ({
    status: 'succeeded';
} & SucceededJobStatus) | ({
    status: 'partially_succeeded';
} & PartiallySucceededJobStatus) | ({
    status: 'failed';
} & FailedJobStatus);